Create compareTriplets
Alice and Bob each created one problem for HackerRank. A reviewer rates the two challenges, awarding points on a scale from  to  for three categories: problem clarity, originality, and difficulty.
We define the rating for Alice's challenge to be the triplet , and the rating for Bob's challenge to be the triplet
If a[0] > b[0] then Alice is awarded  point.
If a[1] > b[1]  then Bob is awarded  point.
If a[2] == b[2] then neither person receives a point.
Comparison points is the total points a person earned.

For example a[1,2,3] and b[3,2,1]. For elements a[0] & b[0] Bob is awarded a point because 3 > 1. For the equal elements a[1] and b[1], no points are earned. Finally, for elements a[2], b[2] so Alice receives a point. Your return array would be [1,1] with Alice's score first and Bob's second.

compareTriplets has the following parameter(s):
a: an array of integers representing Alice's challenge rating
b: an array of integers representing Bob's challenge rating

Output Format
Return an array of two integers denoting the respective comparison points earned by Alice and Bob.
5 6 7
3 6 10
Sample Output:
1 1